Latent_Diffusion_Generator/Outpainting&2d-Imputation/Train/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.25.0
diffusers>=0.25.0
transformers>=4.36.0

numpy>=1.23.0
pillow>=10.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
pyyaml>=6.0.0
